refactor(supply): split native_supply_methods.rs into focused submodules
native_supply_methods.rs had grown to 3671 lines (7x the 500-line guideline).
Split its `impl Interpreter` methods into cohesive submodules of `runtime`,
leaving only the shared helpers and the `QuitOutcome` enum behind:

- native_supply_dispatch.rs    — `native_supply` (the immutable dispatch)
- native_supplier_methods.rs   — `native_supplier` / `native_supplier_mut`
- native_supply_mut_methods.rs — `native_supply_mut` (tap/act)
- supply_classify.rs           — unique / classify
- supply_promise.rs            — on-demand body + `.Promise`/await
- supply_transform.rs          — throttle / from-values / transform / start
- native_supply_methods.rs     — shared markers, quit handling (189 lines)

Pure code move: each method keeps its bytes; only top-level private `fn`
headers were bumped to `pub(super) fn` so cross-module `Self::` calls resolve.
Method count preserved (26 -> 26). No behavior change.

Verified: cargo build / clippy -D warnings / fmt --check clean, make test (7051).
