+++ src/runtime/mod.rs
@@ -192,7 +192,10 @@ mod native_io;
 mod native_io_special;
 pub(crate) mod native_methods;
 mod native_proc_async;
+mod native_supplier_methods;
+mod native_supply_dispatch;
 mod native_supply_methods;
+mod native_supply_mut_methods;
 pub(crate) mod native_types;
 mod ops;
 mod output_sink;
@@ -214,6 +217,9 @@ mod sprintf_helpers;
 mod sprintf_validate;
 pub(crate) mod str_numeric;
 mod subtest;
+mod supply_classify;
+mod supply_promise;
+mod supply_transform;
 mod system;
 mod tap_state;
 mod test_functions;
